Add IPScanParameters
Renamed the old IPScanParameters to Parameters and added new IPScanParameters with permanent properties for only IPScan
